Add add/removeEventListener overloads from HTMLElement
WebviewTag adds electron-specific overloads to addEventListener and
removeEventListener when it extends HTMLElement. However, it does not
include the original overloads from HTMLElement, and Typescript does not
include them when WebviewTag extends HTMLElement. This change adds
those previously-missed overloads to WebviewTag. It is necessary to
allow users of electron.d.ts to compile with Typescript 2.6 --strict.