fix: ensure Electron.Utility property exists (#247)
* fix: ensure Electron.Utility property exists

Added in #246, the Utility namespace currently has no concrete values in
it on electron/electron@main. This change ensures Electron.Utility is
available as a property even when the namespace has no values.

* fix: only set Utility property if namespace empty

If the Utility namespace has concrete values and the const Utility
property is set it causes `error TS2300: Duplicate identifier 'Utility'.`
This tracks whether any values are added to the Utility namespace and
only sets the property workaround if there are none.